Separate structures from connections
